refactor(pipeline): split runtime modules
